refactor: added private function calculateParkingFee Checkout entity and modify parkingFee propertie type

+++ src/main/java/com/pablords/parking/adapters/outbound/database/jpa/models/CheckoutModel.java
@@ -29,5 +29,5 @@ public class CheckoutModel {
     private LocalDateTime checkOutTime;
 
     @Column(name = "parkingFee")
-    private long parkingFee;
+    private Double parkingFee;
 }
